aptransco-backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.1
sqlalchemy>=2.0.30
asyncpg>=0.29.0
alembic>=1.13.1
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
httpx>=0.27.0
python-multipart>=0.0.9
pandas>=2.2.2
openpyxl>=3.1.4
python-dotenv>=1.0.1
psycopg2-binary>=2.9.9
aiofiles>=23.2.1
email-validator>=2.1.1
